% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddirmnom.R
\name{ddirmnom}
\alias{ddirmnom}
\title{calculates the PDF of a dirichlet-multinomial compound distribution for the data x}
\usage{
ddirmnom(x, alpha, log = FALSE)
}
\arguments{
\item{x}{The}

\item{alpha}{The}

\item{log}{Weather}
}
\value{
The probability density value.
}
\description{
calculates the PDF of a dirichlet-multinomial compound distribution for the data x
}
