% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirmnom_fit.R
\name{dirmnom_fit}
\alias{dirmnom_fit}
\title{calculates parameter-optimized fit (including LL) using a dirichlet multinomial model given a matrix of data (k), a model matrix with the features (mm), and a concentration parameter.}
\usage{
dirmnom_fit(k, mm, conc)
}
\arguments{
\item{k}{A matrix with the data to fit a dirichlet multinomial model.}

\item{mm}{A model matrix with the features.}

\item{conc}{The concentration parameter that accounts for overdispersion. To estimate this value use the estimate_overdispersion() function.}
}
\value{
A list with beta = a matrix with the best set of parameters found; and ll = the log-likelihood.
}
\description{
calculates parameter-optimized fit (including LL) using a dirichlet multinomial model given a matrix of data (k), a model matrix with the features (mm), and a concentration parameter.
}
