% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overdispersion.R
\name{estimate_overdispersion}
\alias{estimate_overdispersion}
\title{Estimates overdispersion in multinomial models}
\usage{
estimate_overdispersion(
  y,
  fitted_values,
  coefs,
  method = "Afroz",
  number_samples = NULL,
  weight = NULL
)
}
\arguments{
\item{y}{The vector of multinomial observations.}

\item{fitted_values}{The fitted mean values.}

\item{coefs}{The Coefficients or the number of coefficients.}

\item{method}{The method of estimating overdispersion. One of "Afroz", "Pearson", "Fletcher", or "Deviance". "Afroz" is recommended.}

\item{number_samples}{The number of samples that were fitted. Usually the number of rows in the matrix fitted (\code{y}). Only necessary if providing a flattened \code{y}.}

\item{weight}{TODO}
}
\value{
The estimation for overdispersion.
}
\description{
This is an adaptation of code found spread out in the package mclogit.
The methods for estimating overdispersion are described in Afroz 2018. Summarized in page 54 and 55.
}
