% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirmnom_ll.R
\name{rough_dirmnom_ll}
\alias{rough_dirmnom_ll}
\title{calculates a parameter-non-optimized log-likelihood using a dirichlet multinomial model given a matrix of data (k), a model matrix with the features (mm), and a concentration parameter.
This is useful for data sets that fail to converge the optimization process. e.g. those with not so many samples.}
\usage{
rough_dirmnom_ll(k, mm, conc)
}
\arguments{
\item{k}{A matrix with the data to fit a dirichlet multinomial model.}

\item{mm}{A model matrix with the features.}

\item{conc}{The concentration parameter that accounts for overdispersion. To estimate this value use the estimate_overdispersion() function.}
}
\value{
The log-likelihood for the fit.
}
\description{
calculates a parameter-non-optimized log-likelihood using a dirichlet multinomial model given a matrix of data (k), a model matrix with the features (mm), and a concentration parameter.
This is useful for data sets that fail to converge the optimization process. e.g. those with not so many samples.
}
